+++ tests/integration/utils.rs
@@ -13,6 +13,10 @@ const RECONNECT_DELAY: u32 = 500;
 #[cfg(not(feature = "chaos-monkey"))]
 const RECONNECT_DELAY: u32 = 1000;
 
+pub fn read_env_var(name: &str) -> Option<String> {
+  env::var_os(name).and_then(|s| s.into_string().ok())
+}
+
 fn read_fail_fast_env() -> bool {
   match env::var_os("FRED_FAIL_FAST") {
     Some(s) => match s.into_string() {
@@ -23,6 +27,25 @@ fn read_fail_fast_env() -> bool {
       Err(_) => true,
     },
     None => true,
+  }
+}
+
+#[cfg(feature = "sentinel-auth")]
+fn read_redis_password() -> String {
+  read_env_var("REDIS_PASSWORD").expect("Failed to read REDIS_PASSWORD env")
+}
+
+#[cfg(feature = "sentinel-auth")]
+fn read_sentinel_password() -> String {
+  read_env_var("REDIS_SENTINEL_PASSWORD").expect("Failed to read REDIS_SENTINEL_PASSWORD env")
+}
+
+#[cfg(feature = "sentinel-tests")]
+fn read_sentinel_hostname() -> String {
+  if read_env_var("CIRCLECI_TESTS").is_some() {
+    "redis-sentinel-1".to_owned()
+  }else{
+    "127.0.0.1".to_owned()
   }
 }
 
@@ -39,17 +62,17 @@ where
     fail_fast: read_fail_fast_env(),
     server: ServerConfig::Sentinel {
       hosts: vec![
-        ("127.0.0.1".into(), 26379),
-        ("127.0.0.1".into(), 26380),
-        ("127.0.0.1".into(), 26381),
+        (read_sentinel_hostname(), 26379),
+        (read_sentinel_hostname(), 26380),
+        (read_sentinel_hostname(), 26381),
       ],
       service_name: "redis-sentinel-main".into(),
-      #[cfg(feature = "sentinel-auth")]
+      // TODO fix this so sentinel-tests can run without sentinel-auth
       username: None,
-      #[cfg(feature = "sentinel-auth")]
-      password: None,
+      password: Some(read_sentinel_password()),
     },
     pipeline,
+    password: Some(read_redis_password()),
     ..Default::default()
   };
   let client = RedisClient::new(config.clone());