+++ tests/runners/no-features.sh
@@ -1,0 +1,13 @@
+#!/bin/bash
+
+declare -a arr=("REDIS_VERSION" "REDIS_USERNAME" "REDIS_PASSWORD" "REDIS_SENTINEL_PASSWORD")
+
+for env in "${arr[@]}"
+do
+  if [ -z "$env" ]; then
+    echo "$env must be set. Run `source tests/environ` if needed."
+    exit 1
+  fi
+done
+
+cargo test --release --lib --tests --no-default-features -- --test-threads=1 "$@"